% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_density_data}
\alias{compute_density_data}
\title{Compute Density Data for Plotting}
\usage{
compute_density_data(
  data,
  variable_col,
  group_by_treatment = FALSE,
  q1_quantile = 0.25,
  q3_quantile = 0.75
)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{variable_col}{Name of the column for which to compute density statistics.}

\item{group_by_treatment}{Logical, whether to group by the "Treatment" column. Default is FALSE.}

\item{q1_quantile}{Quantile for the lower bound. Default is 0.25.}

\item{q3_quantile}{Quantile for the upper bound. Default is 0.75.}
}
\value{
A dataframe containing density data, mean, and quantile y-values.
}
\description{
Calculates density data, mean, and quantile information for a specified column
in a dataframe, which can be used for plotting density plots with additional statistics.
}
