% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_df_user.R
\name{load_df_user}
\alias{load_df_user}
\title{Load User Reduction data}
\usage{
load_df_user(excel_path)
}
\arguments{
\item{excel_path}{Path to the excel  file}
}
\value{
List of cleaned data frames for analysis
}
\description{
This function takes in the excel path and creates a list of data frames from
the sheet names and updates columns to have correct types
}
