% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapDamage_param.R
\name{plot_mapDamage_param}
\alias{plot_mapDamage_param}
\title{Plotting mapDamage parameters used to characterize ancient DNA}
\usage{
plot_mapDamage_param(mapdam_params, parameter, metric = "Mean")
}
\arguments{
\item{mapdam_params}{df}

\item{parameter}{parameter to plot}

\item{metric}{using mean metric}
}
\value{

}
\description{
Plots the 4 damage parameters used for the bayesian statistics for determining
the aDNA damage characterisitics.
}
