% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_group.R
\name{plot_qc_group}
\alias{plot_qc_group}
\title{Plot Quality-Control density plots}
\usage{
plot_qc_group(
  data_frame_user,
  qc_sheet = "Statistics",
  variable_col,
  treatment_group = TRUE
)
}
\arguments{
\item{data_frame_user}{project data frame}

\item{qc_sheet}{Name of dataframe from project_dataframe. Default value = "Statistics".}

\item{variable_col}{Name of summary statistic for plotting.}

\item{treatment_group}{Logical, whether to group by the "Treatment" column. Default is TRUE.}
}
\value{
Single ggplot object. Distribution of selected summary statistic for Quality control
}
\description{
This function generates a density plot for quality control visualization after NGS,
prior to downstream analysis grouped by Treatment. It can optionally plot statistical information such as
mean and quantiles.
}
