% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raincloud.R
\name{plot_raincloud}
\alias{plot_raincloud}
\title{Plots point,density and boxplots side-by-side per treatment}
\usage{
plot_raincloud(user_dataframe, observation_name, treatment = "Treatment", ...)
}
\arguments{
\item{user_dataframe}{data frame for plotting}

\item{observation_name}{Statistical variable for plotting}

\item{treatment}{group-by variable for ggplot aesthetics}

\item{...}{}
}
\value{
ggplot object
}
\description{
Plots point,density and boxplots side-by-side per treatment
}
\examples{
raincloud_plot(data_frame, "ReadLenTrim")
}
