% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_USER.R
\docType{data}
\name{var_factors}
\alias{var_factors}
\title{Clean USER data}
\format{
An object of class \code{character} of length 4.
}
\usage{
var_factors
}
\arguments{
\item{dataframe}{df containing data from USER treatment}
}
\value{
a cleaned df for plotting
}
\description{
Used to clean and prepare USER treatment libraries after sequencing and basic
statistic summaries
}
\keyword{datasets}
